pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "times-data"
version = "0.1.0"
description = "Open-source data layer for TIMES energy models. Import, validate, edit, and compile in Python."
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.4",
]

[project.scripts]
times-data = "times_data.cli.main:cli"

[tool.setuptools.packages.find]
include = ["times_data*"]

[tool.setuptools.package-data]
"times_data.wiki" = ["data/**/*.md", "data/**/*.yaml", "data/**/*.yml", "data/**/*.json"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests", "interface/backend/tests"]
